+++ Dorset_OOP_Project/Payment.cs
@@ -5,25 +5,25 @@ namespace Dorset_OOP_Project
     {
         public int PaymentId { get; set; }
         public int InvoiceId { get; set; }
-        public float Amount { get; set; }
+        public double Amount { get; set; }
         public DateTime Date { get; set; }
         public string Method { get; set; }
 
-        public Payment(float amount, string method)
+        public Payment(double amount, string method)
         {
             Amount = amount;
             Date = DateTime.Now;
             Method = method;
         }
 
-        public Payment(float amount, DateTime date, string method)
+        public Payment(double amount, DateTime date, string method)
         {
             Amount = amount;
             Date = date;
             Method = method;
         }
 
-        public Payment(float amount, DateTime date, string method, int invoiceId, int paymentId)
+        public Payment(double amount, DateTime date, string method, int invoiceId, int paymentId)
         {
             Amount = amount;
             Date = date;
